refactor: move Format.kt to util
